feat: SFORM-4119 123

+++ src/utils/jiraAccount.js
@@ -1,9 +1,22 @@
-// const puppeteer = require('puppeteer');
-// const puppeteer = require('puppeteer');
 import puppeteer from 'puppeteer'
 import path from 'path'
+import { BoardBug }from './constants'
 
 const getListUrl = 'http://jira.taimei.com/rest/greenhopper/1.0/xboard/work/allData.json'
+
+const questionUrl = 'http://jira.taimei.com/rest/issueNav/1/issueTable'
+
+async function toQuestionPage (page) {
+    // 得到问题按钮
+    const questionBtn = await page.$('#find_link');
+    // 点击获取下拉
+    await questionBtn.click();
+    // 等他渲染 来个1秒钟
+    await page.waitForTimeout(1000)
+    // 获取面板下拉数据   ----- 面板
+    const linkToBtn = await page.$('#filter_lnk_my_lnk')
+    linkToBtn.click()
+}
 
 export const initAccount = async function (account) {
     return new Promise(async (resolve, reject) => {
@@ -16,8 +29,22 @@ export const initAccount = async function (account) {
             }
         })
         try {
+            let questionData = {}
             // 新建界面
             const page = await browser.newPage()
+            await page.setRequestInterception(true);
+            page.on('request', async req => {
+                req.continue({})
+            })
+            page.on('response', async res => {
+                if (res.url().indexOf(questionUrl) !== -1) {
+                    console.log('哈哈哈')
+                    questionData = await res.json()
+                    await browser.close();
+                    resolve(questionData)
+                }
+                return res
+            });
             // 跳转界面
             await page.goto('http://jira.taimei.com/login.jsp')
             const loginInput = await page.$('#login-form-username');
@@ -36,13 +63,20 @@ export const initAccount = async function (account) {
             await page.waitForTimeout(account.delay)
             // 得到看板按钮
             const boardBtn = await page.$('#greenhopper_menu');
-            // 点击按钮
             await page.screenshot({
                path: path.resolve(__dirname, '../account.png')
             })
+            if (account.designatedBoard.trim().toLocaleLowerCase() === BoardBug) {
+                // 我未完成的问题
+                await toQuestionPage(page)
+                return 
+            } 
+            // 点击按钮-看板的
             await boardBtn.click();
-            // 获取下拉数据
+            // 获取面板下拉数据   ----- 面板
             const boardList = []
+            // 等待2秒 跳转需要时间
+            await page.waitForTimeout(1000)
             const linkList = await page.$$('#greenhopper_menu_dropdown_recent .aui-list-truncate li');
             for (var i = 0; i< linkList.length; i++) {
                 let data = await linkList[i].$eval('a', el => {
@@ -56,6 +90,7 @@ export const initAccount = async function (account) {
                 })
                 boardList.push(data)
             }
+            console.log('boardList', boardList, linkList)
             const hasBoard = boardList.filter((item) => {
                 const text = item.innerHTML;
                 if (text.trim() === account.designatedBoard.trim()) {
